Introduce test subcommand for debug builds
